End of section eighteen

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,7 +16,8 @@ export class MessagesComponent implements OnInit {
   pageSize:number;
   container:string='Inbox';
   loading:boolean = false;
-  constructor(private messageService : MessageService) { }
+  constructor(private messageService : MessageService,
+    private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -38,8 +40,13 @@ export class MessagesComponent implements OnInit {
 
   }
   onDeleteMessage(message: Message){
-    this.messageService.deleteMessage(message.id).subscribe(() => {
-      this.messages = this.messages.filter(m => m.id != message.id);
-    })
+    this.confirmService.openModalWithComponent("Deleting message").subscribe(result => {
+      if(result){
+        this.messageService.deleteMessage(message.id).subscribe(() => {
+          this.messages = this.messages.filter(m => m.id != message.id);
+        })
+      }
+    });
+    
   }
 }
